requirements.txt
slack-bolt>=1.16.0
python-dotenv>=0.21.0
flask
websocket-client>=1.3.3
gspread>=5.7.0
oauth2client>=4.1.3 
